requirements.txt
pandas
numpy
openpyxl
matplotlib
pillow
tabulate